backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
pydantic>=2.5.3
pydantic-settings>=2.1.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
pyodbc>=5.0.1
werkzeug>=3.0.0